+++ mod.js
@@ -3,4 +3,4 @@ import Change from "./src/Change.js";
 import Changelog from "./src/Changelog.js";
 import Release from "./src/Release.js";
 
-export { parser, Change, Changelog, Release };
+export { Change, Changelog, parser, Release };