+++ src/deps.js
@@ -1,1 +1,4 @@
-export { eq, default as Semver } from "https://deno.land/x/semver@v1.3.0/mod.ts";
+export {
+  default as Semver,
+  eq,
+} from "https://deno.land/x/semver@v1.3.0/mod.ts";