+++ CHANGELOG.md
@@ -1,9 +1,16 @@
+<!-- deno-fmt-ignore-file -->
+
 # Changelog
 
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## [1.1.0] - 2021-05-13
+### Added
+- Support for flags like 'deno-fmt-ignore-file'
+- Allow to pass a version number to `--release` command [#18]
 
 ## [1.0.1] - 2020-08-31
 ### Fixed
@@ -167,7 +174,9 @@ First version
 [#10]: https://github.com/oscarotero/keep-a-changelog/issues/10
 [#12]: https://github.com/oscarotero/keep-a-changelog/issues/12
 [#13]: https://github.com/oscarotero/keep-a-changelog/issues/13
+[#18]: https://github.com/oscarotero/keep-a-changelog/issues/18
 
+[1.1.0]: https://github.com/oscarotero/keep-a-changelog/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/oscarotero/keep-a-changelog/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/oscarotero/keep-a-changelog/compare/v0.10.1...v1.0.0
 [0.10.1]: https://github.com/oscarotero/keep-a-changelog/compare/v0.10.0...v0.10.1